+++ src/Dots.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import { space, color } from 'styled-system'
 import Flex from './Flex'
@@ -22,6 +23,10 @@ const Dot = styled.button([], {
   space,
   color
 )
+Dot.propTypes = {
+  ...space.propTypes,
+  ...color.propTypes
+}
 Dot.defaultProps = {
   m: 0,
   p: 1,
@@ -46,5 +51,11 @@ export const Dots = ({
       />
     ))}
   </Flex>
+
+Dots.propTypes = {
+  index: PropTypes.number.isRequired,
+  length: PropTypes.number.isRequired,
+  onClick: PropTypes.func
+}
 
 export default Dots