+++ src/Presenter.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import Box from './Box'
 import Flex from './Flex'
 import Zoom from './Zoom'
@@ -13,6 +14,8 @@ export const Presenter = ({
   slides = [],
   mode,
   notes = {},
+  update,
+  step,
   ...props
 }) => {
   const Next = slides[index + 1]
@@ -53,7 +56,7 @@ export const Presenter = ({
             <Zoom zoom={1/4}>
               <Root {...props}>
                 {Next && (
-                  <Slide>
+                  <Slide index={index + 1}>
                     <Next />
                   </Slide>
                 )}
@@ -76,6 +79,16 @@ export const Presenter = ({
       </Flex>
     </Flex>
   )
+}
+
+Presenter.propTypes = {
+  index: PropTypes.number.isRequired,
+  length: PropTypes.number.isRequired,
+  update: PropTypes.func.isRequired,
+  step: PropTypes.number.isRequired,
+  slides: PropTypes.array,
+  mode: PropTypes.string,
+  notes: PropTypes.object
 }
 
 export default Presenter